add activities, fragments, animations and fix bugs on Ruang Bangun

+++ app/src/main/java/polinema/ac/id/rumusbangun/fragments/Bola/VolumeBola.java
@@ -44,7 +44,7 @@ public class VolumeBola extends Fragment {
                 if(mListener != null){
                     String strjari = jari.getText().toString();
                     if(!TextUtils.isEmpty(strjari)){
-                        int intJari = Integer.parseInt(strjari);
+                        float intJari = (float) Double.parseDouble(strjari);
                         rumus r = new rumus();
                         r.vBola(intJari);
                         mListener.hitungVolumeBola(r.getHasil());
