+++ app/src/main/java/polinema/ac/id/rumusbangun/fragments/Tabung/LuasAlasTabung.java
@@ -44,7 +44,7 @@ public class LuasAlasTabung extends Fragment {
                 if(mListener != null){
                     String strjari = jari.getText().toString();
                     if(!TextUtils.isEmpty(strjari)){
-                        int intJari = Integer.parseInt(strjari);
+                        float intJari = (float) Double.parseDouble(strjari);
                         rumus r = new rumus();
                         r.laTabung(intJari);
                         mListener.hitungLuasAlasTabung(r.getHasil());
